Fixes on REST fields
